feat: half-star book ratings (0.5 increments, e.g. 4.5, 3.5)
Migration 0034: ALTER rating column from SMALLINT to NUMERIC(2,1).
Model: Rating field changed from int to float64.
Controller: ParseFloat instead of Atoi for rating form value.

New ratingStars template function renders full/half/empty stars using
CSS clip (left 50% overflow:hidden) for the half-star effect. Replaces
all inline star-rating logic across 4 templates.

Editor: replaced 5 radio buttons with a <select> dropdown (0.5 steps
from 0 to 5) + live star preview that updates on change.

+++ models/book.go
@@ -34,7 +34,7 @@ type Book struct {
 	MyThoughtsHTML    template.HTML
 	LinkURL           string
 	ReadingStatus     string
-	Rating            int
+	Rating            float64
 	DateStarted       string // YYYY-MM-DD or empty
 	DateFinished      string // YYYY-MM-DD or empty
 	IsPublished       bool
@@ -101,7 +101,7 @@ func parseDateToNullTime(s string) sql.NullTime {
 }
 
 // Create creates a new book and returns it.
-func (bs *BookService) Create(userID int, title, slug, bookAuthor, isbn, publisher string, pageCount int, coverImageURL, content, description, myThoughts, linkURL, readingStatus string, rating int, dateStarted, dateFinished string, isPublished bool, genreIDs []int) (*Book, error) {
+func (bs *BookService) Create(userID int, title, slug, bookAuthor, isbn, publisher string, pageCount int, coverImageURL, content, description, myThoughts, linkURL, readingStatus string, rating float64, dateStarted, dateFinished string, isPublished bool, genreIDs []int) (*Book, error) {
 	if slug == "" {
 		slug = generateSlug(title)
 	} else {
@@ -153,7 +153,7 @@ func (bs *BookService) Create(userID int, title, slug, bookAuthor, isbn, publish
 }
 
 // Update updates an existing book.
-func (bs *BookService) Update(bookID int, title, slug, bookAuthor, isbn, publisher string, pageCount int, coverImageURL, content, description, myThoughts, linkURL, readingStatus string, rating int, dateStarted, dateFinished string, isPublished bool, genreIDs []int) error {
+func (bs *BookService) Update(bookID int, title, slug, bookAuthor, isbn, publisher string, pageCount int, coverImageURL, content, description, myThoughts, linkURL, readingStatus string, rating float64, dateStarted, dateFinished string, isPublished bool, genreIDs []int) error {
 	if slug == "" {
 		slug = generateSlug(title)
 	} else {
